tests: Update and add tests for RFC 2203.
