+++ src/test/ui/consts/rfc-2203-const-array-repeat-exprs/migrate-pass.rs
@@ -1,6 +1,8 @@
-// ignore-compile-mode-nll
+// check-pass
+// compile-flags: -Z borrowck=migrate
+// ignore-compare-mode-nll
+#![feature(const_in_array_repeat_expressions)]
 #![allow(warnings)]
-#![feature(constants_in_array_repeat_expressions, nll)]
 
 // Some type that is not copyable.
 struct Bar;
@@ -82,12 +84,6 @@ mod non_constants {
         let arr: [Option<Bar>; 1] = [x; 1];
     }
 
-    fn no_impl_copy_empty_value_multiple_elements() {
-        let x = None;
-        let arr: [Option<Bar>; 2] = [x; 2];
-        //~^ ERROR the trait bound `std::option::Option<Bar>: std::marker::Copy` is not satisfied [E0277]
-    }
-
     fn no_impl_copy_value_no_elements() {
         let x = Some(Bar);
         let arr: [Option<Bar>; 0] = [x; 0];
@@ -96,12 +92,6 @@ mod non_constants {
     fn no_impl_copy_value_single_element() {
         let x = Some(Bar);
         let arr: [Option<Bar>; 1] = [x; 1];
-    }
-
-    fn no_impl_copy_value_multiple_elements() {
-        let x = Some(Bar);
-        let arr: [Option<Bar>; 2] = [x; 2];
-        //~^ ERROR the trait bound `std::option::Option<Bar>: std::marker::Copy` is not satisfied [E0277]
     }
 
     fn impl_copy_empty_value_no_elements() {
