+++ src/test/ui/consts/rfc-2203-const-array-repeat-exprs/nll-pass.rs
@@ -1,7 +1,7 @@
-// ignore-compile-mode-nll
-// compile-flags: -Z borrowck=migrate
-#![feature(constants_in_array_repeat_expressions)]
+// check-pass
+// ignore-compare-mode-nll
 #![allow(warnings)]
+#![feature(const_in_array_repeat_expressions, nll)]
 
 // Some type that is not copyable.
 struct Bar;
@@ -83,12 +83,6 @@ mod non_constants {
         let arr: [Option<Bar>; 1] = [x; 1];
     }
 
-    fn no_impl_copy_empty_value_multiple_elements() {
-        let x = None;
-        let arr: [Option<Bar>; 2] = [x; 2];
-        //~^ ERROR the trait bound `std::option::Option<Bar>: std::marker::Copy` is not satisfied [E0277]
-    }
-
     fn no_impl_copy_value_no_elements() {
         let x = Some(Bar);
         let arr: [Option<Bar>; 0] = [x; 0];
@@ -97,12 +91,6 @@ mod non_constants {
     fn no_impl_copy_value_single_element() {
         let x = Some(Bar);
         let arr: [Option<Bar>; 1] = [x; 1];
-    }
-
-    fn no_impl_copy_value_multiple_elements() {
-        let x = Some(Bar);
-        let arr: [Option<Bar>; 2] = [x; 2];
-        //~^ ERROR the trait bound `std::option::Option<Bar>: std::marker::Copy` is not satisfied [E0277]
     }
 
     fn impl_copy_empty_value_no_elements() {
